removing submenus on therapy tabs OP-492
