+++ src/components/accessories/patientTherapy/PatientTherapy.tsx
@@ -1,0 +1,12 @@
+import React, { FC } from "react";
+import TherapyForm from "./therapyForm/TherapyForm";
+
+const PatientTherapy: FC = () => {
+    return (
+        <div className="patientTherapy">
+            <TherapyForm/>
+        </div>
+    )
+}
+
+export default PatientTherapy
