--- src/components/accessories/patientTherapy/manageTherapy/ManagePatientTherapy.tsx
@@ -1,11 +1,0 @@
-import React, { FC } from "react";
-
-const ManagePatientTherapy: FC = () => {
-    return (
-        <div className="patientTherapy">
-            WELCOME TO MANAGE THERAPY
-        </div>
-    )
-}
-
-export default ManagePatientTherapy
