--- src/components/accessories/patientTherapy/newTherapy/NewPatientTherapy.tsx
@@ -1,12 +1,0 @@
-import React, { FC } from "react";
-import NewTherapyForm from "./newTherapyForm/NewTherapyForm";
-
-const NewPatientTherapy: FC = () => {
-    return (
-        <div className="patientTherapy">
-            <NewTherapyForm/>
-        </div>
-    )
-}
-
-export default NewPatientTherapy
