+++ src/components/accessories/patientTherapy/therapyForm/styles.scss
@@ -1,0 +1,71 @@
+@import "../../../../styles/variables";
+@import "../../../../../node_modules/susy/sass/susy";
+
+.patientTherapyForm {
+    display: inline-block;
+    flex-direction: column;
+    align-items: center;
+    width: 100%;
+
+    .patientTherapyForm__item {
+        margin: 7px 0;
+        padding: 0 15px;
+        width: 33%;
+        @include susy-media($narrow) {
+            padding: 0 10px;
+        }
+        @include susy-media($tablet_land) {
+            padding: 0 10px;
+        }
+        @include susy-media($tablet_port) {
+            width: 50%;
+        }
+        @include susy-media($smartphone) {
+            width: 100%;
+        }
+        .dateField,
+        .textField,
+        .selectField {
+            width: 100%;
+        }
+
+        &.fullWidth {
+            width: 100%;
+        }
+
+        &.compressed {
+            .textField {
+                float: left;
+                width: 50%;
+                &:nth-of-type(1) {
+                    padding-right: 10px;
+                }
+            }
+        }
+    }
+
+    .patientTherapyForm__buttonSet {
+        display: flex;
+        margin-top: 25px;
+        padding: 0 15px;
+        flex-direction: row-reverse;
+        @include susy-media($smartphone_small) {
+            display: block;
+        }
+
+        .submit_button,
+        .reset_button {
+            .MuiButton-label {
+                font-size: smaller;
+                letter-spacing: 1px;
+                font-weight: 600;
+            }
+            button {
+                @include susy-media($smartphone_small) {
+                    width: 100%;
+                    margin-top: 10px;
+                }
+            }
+        }
+    }
+}
