+++ src/components/accessories/patientTriage/patientTriageForm/types.ts
@@ -28,4 +28,7 @@ export type TPatientTriageFormFieldName =
   | "diuresis"
   | "bowel"
   | "auscultation";
+
+
+
