+++ src/components/accessories/tabs/RouterTabs.tsx
@@ -5,20 +5,16 @@ import { Switch } from "react-router";
 import { IProps } from "./types";
 import PrivateRoute from "../privateRoute/PrivateRoute";
 import "./styles.scss";
-import ExpandLessIcon from '@material-ui/icons/ExpandLess';
-import ExpandMoreIcon from '@material-ui/icons/ExpandMore';
 
 const RouterTabs: FunctionComponent<IProps> = ({ config, defaultRoute }) => {
   const history = useHistory();
   const match = useRouteMatch();
   const { url } = match;
   const {pathname} = useLocation();
-  const [showMenu, setShowMenu] = useState(false);
   const currentPath: string | undefined = config.map(item => item.path).find(path => pathname.match(new RegExp(`^(${url}${path})([/?].*)?$`, "gi")));
 
   const switchTab = (path: string) => {
     history.push(url + path);
-    if(showMenu) setShowMenu(false);
   };
 
   useEffect(() => {
@@ -35,23 +31,8 @@ const RouterTabs: FunctionComponent<IProps> = ({ config, defaultRoute }) => {
         const path = item.path!;
         return (
 
-          <div className={classNames("tab", { active: (path === currentPath) })} key={index} onClick={() => !item.submenu &&  switchTab(path)}>
-            <span>{item.label}</span>
-           
-              {item.submenu &&
-                <ul className="menu">
-                  {!showMenu && <ExpandMoreIcon onClick={() => setShowMenu(true)} /> }
-                  {showMenu && <ExpandLessIcon onClick={() => setShowMenu(false)} />}            
-                  {showMenu && item.submenu?.map((subitem, sindex) =>{
-                    return ( 
-                      <li key={sindex} onClick={() => switchTab(subitem.path!)}>
-                          {subitem.label}
-                      </li>
-                    )
-                  })}
-                </ul>
-              }
-            
+          <div className={classNames("tab", { active: (path === currentPath) })} key={index} onClick={() => switchTab(path)}>
+            <span>{item.label}</span>            
           </div>
         )
       }) 
@@ -61,17 +42,6 @@ const RouterTabs: FunctionComponent<IProps> = ({ config, defaultRoute }) => {
         return (
           <option className={classNames("tab", { active: (path === currentPath) })} key={index} value={path}>
             {item.label}
-            {item.submenu &&
-                <ul>
-                  {item.submenu?.map((subitem, sindex) =>{
-                    return ( 
-                      <li key={sindex} onClick={() => switchTab(subitem.path!)}>
-                          {subitem.label}
-                      </li>
-                    )
-                  })}
-                </ul>
-            }
           </option>
         )
       })
@@ -93,23 +63,13 @@ const RouterTabs: FunctionComponent<IProps> = ({ config, defaultRoute }) => {
       <div className="tabs_content"> 
         <Switch>
           {config.map((item, index) => (
-                   
-            !item.submenu
-            ?              
+                            
              <PrivateRoute path={url + item.path} key={index}>
                 <div className="panel">
                   {item.content}
                 </div>
               </PrivateRoute>
-            :
-            item.submenu.map((subitem, suindex) => (
-              <PrivateRoute path={url + subitem.path} key={suindex}>
-                <div className="panel">
-                  {subitem.content}
-                </div>
-              </PrivateRoute>
-            ))
-          ))}
+           ))}
         </Switch>
       </div>
     </Fragment>
